ref(ui): Allow custom usage chart colors
