+++ static/app/types/core.tsx
@@ -73,6 +73,15 @@ export enum DataCategory {
 
 export type EventType = 'error' | 'transaction' | 'attachment';
 
+export enum Outcome {
+  ACCEPTED = 'accepted',
+  FILTERED = 'filtered',
+  INVALID = 'invalid',
+  DROPPED = 'dropped', // this is not a real outcome coming from the server
+  RATE_LIMITED = 'rate_limited',
+  CLIENT_DISCARD = 'client_discard',
+}
+
 export type IntervalPeriod = ReturnType<typeof getInterval>;
 
 /**
