+++ static/app/views/organizationStats/types.tsx
@@ -1,14 +1,5 @@
 import {SeriesApi} from 'sentry/types';
 
-export enum Outcome {
-  ACCEPTED = 'accepted',
-  FILTERED = 'filtered',
-  INVALID = 'invalid',
-  DROPPED = 'dropped', // this is not a real outcome coming from the server
-  RATE_LIMITED = 'rate_limited',
-  CLIENT_DISCARD = 'client_discard',
-}
-
 /**
  * Raw response from API endpoint
  */
