+++ static/app/views/organizationStats/usageStatsPerMin.tsx
@@ -2,9 +2,9 @@ import styled from '@emotion/styled';
 
 import AsyncComponent from 'sentry/components/asyncComponent';
 import {t} from 'sentry/locale';
-import {DataCategory, Organization} from 'sentry/types';
+import {DataCategory, Organization, Outcome} from 'sentry/types';
 
-import {Outcome, UsageSeries} from './types';
+import {UsageSeries} from './types';
 import {formatUsageWithUnits, getFormatUsageOptions} from './utils';
 
 type Props = {
