+++ static/app/views/settings/project/server-side-sampling/utils/projectStatsToPredictedSeries.tsx
@@ -1,11 +1,10 @@
 import moment from 'moment';
 
 import {t} from 'sentry/locale';
-import {SeriesApi} from 'sentry/types';
+import {Outcome, SeriesApi} from 'sentry/types';
 import {Series} from 'sentry/types/echarts';
 import {defined} from 'sentry/utils';
 import commonTheme from 'sentry/utils/theme';
-import {Outcome} from 'sentry/views/organizationStats/types';
 
 import {quantityField} from '.';
