+++ static/app/views/settings/project/server-side-sampling/utils/projectStatsToSeries.tsx
@@ -2,11 +2,10 @@ import cloneDeep from 'lodash/cloneDeep';
 import moment from 'moment';
 
 import {t} from 'sentry/locale';
-import {SeriesApi} from 'sentry/types';
+import {Outcome, SeriesApi} from 'sentry/types';
 import {Series} from 'sentry/types/echarts';
 import {defined} from 'sentry/utils';
 import commonTheme from 'sentry/utils/theme';
-import {Outcome} from 'sentry/views/organizationStats/types';
 
 import {quantityField} from '.';
